// src/pages/channels-5_3/interfaces/channel.interface.ts
export interface ChannelInterface {
  _id: string;
  code: string;
  stream_identifier: string;
  international_stream_number: string;
  forward_stream_interaction_number: string;
  reverse_stream_interaction_number: string;
  rcpu_site_id_a: string;
  icm_a_stream: string;
  rcpu_site_id_z: string;
  icm_b_stream: string;
  channel_identifier: string;
  channel_link: string;
  channel_number_in_stream: string;
  point_a_of_channel: string;
  point_b_of_channel: string;
  binding_node_a: string;
  binding_node_b: string;
  last_mile_track_consumer_mu_a: string;
  last_mile_track_consumer_mu_b: string;
  channel_consumer: string;
  connection_number: string;
  destination_index: string;
  channel_organization_order_number: string;
  channel_decommission_order_number: string;
  channel_reserve: string;
  channel_mode: string;
  timeslot_number: string;
  fac_group_bc: string;
  site_a: string;
  ne_a_id: string;
  a_port_nms: string;
  site_z: string;
  ne_z_id: string;
  z_port_nms: string;
  order_number: string;
  stream_consumer: string;
  fac_protection_mode: string;
  additional_information: string;
  additional_consumer_information: string;
  phone_numbers: string;
  consumer_2: string;
  consumer_3: string;
  payment_status: string;
  responsibility_zone: string;
  mux_belonging_site_a: string;
  mux_belonging_site_b: string;
  verification_by_letters_from_branches: string;
  verification_result: string;
  verification: string;
  creatorId?: string;
  created_at?: string;
  updated_at?: string;
  __v?: number;
}